feat: update google analytics components
