+++ pages/_app.js
@@ -1,12 +1,13 @@
-import Router from "next/router";
-
-import * as gtag from "../lib/gtag";
+import { GoogleAnalytics } from "../components/GoogleAnalytics";
 
 import "../assets/style.scss";
 
-Router.events.on("routeChangeComplete", (url) => gtag.pageview(url));
-
 // This default export is required in a new `pages/_app.js` file.
 export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  return (
+    <>
+      <GoogleAnalytics />
+      <Component {...pageProps} />
+    </>
+  );
 }
